data-service/app/requirements.txt
fastapi
uvicorn
psycopg2-binary
python-jose
requests
pydantic
sqlalchemy
alembic
python-multipart

